Finished breaking up `functions` index file
