+++ packages/functions/src/index.ts
@@ -4,9 +4,7 @@ import {defineString, projectID} from 'firebase-functions/params';
 import {auth} from 'firebase-functions/v1';
 import {onInit} from 'firebase-functions/v2/core';
 import {onDocumentCreated, onDocumentUpdated} from 'firebase-functions/v2/firestore';
-import {HttpsError, onCall, onRequest} from 'firebase-functions/v2/https';
-
-import {logger} from '@shared/services/logger.shared';
+import {onCall, onRequest} from 'firebase-functions/v2/https';
 
 import {
   ACCOUNTS_DB_COLLECTION,
@@ -16,7 +14,6 @@ import {
   IMPORT_QUEUE_DB_COLLECTION,
   USER_FEED_SUBSCRIPTIONS_DB_COLLECTION,
 } from '@shared/lib/constants.shared';
-import {prefixError} from '@shared/lib/errorUtils.shared';
 
 import {parseAccount, parseAccountId, toStorageAccount} from '@shared/parsers/accounts.parser';
 import {parseFeedItem, parseFeedItemId, toStorageFeedItem} from '@shared/parsers/feedItems.parser';
@@ -52,6 +49,10 @@ import {SuperfeedrService} from '@sharedServer/services/superfeedr.server';
 import {ServerUserFeedSubscriptionsService} from '@sharedServer/services/userFeedSubscriptions.server';
 import {WipeoutService} from '@sharedServer/services/wipeout.server';
 
+import {wipeoutAccountHelper} from '@src/lib/accountWipeout';
+import {subscribeAccountToFeedHelper} from '@src/lib/feedSubscription';
+import {handleFeedUnsubscribeHelper} from '@src/lib/feedUnsubscribe';
+import {createImportQueueItemHelper} from '@src/lib/importQueueCreation';
 import {processImportQueueItem} from '@src/lib/importQueueProcessor';
 import {handleSuperfeedrWebhookHelper} from '@src/lib/superfeedrWebhook';
 
@@ -68,7 +69,6 @@ let importQueueService: ServerImportQueueService;
 let wipeoutService: WipeoutService;
 let rssFeedService: ServerRssFeedService;
 let feedItemsService: ServerFeedItemsService;
-let servicesInitialized = false;
 
 // Initialize services on startup
 onInit(() => {
@@ -158,13 +158,11 @@ onInit(() => {
     feedSourcesService,
     userFeedSubscriptionsService,
   });
-
-  servicesInitialized = true;
 });
 
 setGlobalOptions({
-  region: 'us-central1', // TODO: This should probably be an environment variable.
-  invoker: 'private', // Only allow authenticated requests to the functions.
+  region: FIREBASE_FUNCTIONS_REGION,
+  invoker: 'private', // Only allow authenticated requests.
 });
 
 /**
@@ -174,17 +172,6 @@ export const handleSuperfeedrWebhook = onRequest(
   // This webhook is server-to-server, so we don't need to worry about CORS.
   {cors: false},
   async (request, response) => {
-    if (!servicesInitialized) {
-      logger.error(
-        new Error('[SUPERFEEDR] Services not initialized before webhook handler was called')
-      );
-      response.status(500).json({
-        success: false,
-        error: 'Internal server error: Services not initialized',
-      });
-      return;
-    }
-
     await handleSuperfeedrWebhookHelper({
       request,
       response,
@@ -201,13 +188,6 @@ export const handleSuperfeedrWebhook = onRequest(
 export const processImportQueueOnDocumentCreated = onDocumentCreated(
   `/${IMPORT_QUEUE_DB_COLLECTION}/{importQueueItemId}`,
   async (event) => {
-    if (!servicesInitialized) {
-      logger.error(
-        new Error('[IMPORT] Services not initialized before import queue processor was called')
-      );
-      return;
-    }
-
     await processImportQueueItem({
       importQueueItemId: event.params.importQueueItemId,
       data: event.data,
@@ -220,29 +200,10 @@ export const processImportQueueOnDocumentCreated = onDocumentCreated(
  * Permanently deletes all data associated with an account when their Firebase auth user is deleted.
  */
 export const wipeoutAccountOnAuthDelete = auth.user().onDelete(async (firebaseUser) => {
-  logger.log(`[WIPEOUT] Firebase user deleted. Processing account wipeout...`, {
-    fireabseUid: firebaseUser.uid,
+  await wipeoutAccountHelper({
+    firebaseUid: firebaseUser.uid,
+    wipeoutService,
   });
-
-  const accountIdResult = parseAccountId(firebaseUser.uid);
-  if (!accountIdResult.success) {
-    logger.error(
-      prefixError(accountIdResult.error, '[WIPEOUT] Invalid account ID. Not wiping out account.'),
-      {firebaseUid: firebaseUser.uid}
-    );
-    return;
-  }
-  const accountId = accountIdResult.value;
-
-  const wipeoutAccountResult = await wipeoutService.wipeoutAccount(accountId);
-  if (!wipeoutAccountResult.success) {
-    logger.error(prefixError(wipeoutAccountResult.error, '[WIPEOUT] Failed to wipe out account'), {
-      accountId,
-    });
-    return;
-  }
-
-  logger.log(`[WIPEOUT] Successfully wiped out account`, {accountId});
 });
 
 /**
@@ -252,52 +213,11 @@ export const subscribeAccountToFeedOnCall = onCall(
   // TODO: Lock down CORS to only allow requests from my domains.
   {cors: true},
   async (request): Promise<{readonly userFeedSubscriptionId: UserFeedSubscriptionId}> => {
-    if (!request.auth) {
-      // eslint-disable-next-line no-restricted-syntax
-      throw new HttpsError('unauthenticated', 'Must be authenticated');
-    } else if (!request.data.url) {
-      // TODO: Use zod to validate the request data.
-      // eslint-disable-next-line no-restricted-syntax
-      throw new HttpsError('invalid-argument', 'URL is required');
-    }
-
-    const accountIdResult = parseAccountId(request.auth.uid);
-    if (!accountIdResult.success) {
-      // eslint-disable-next-line no-restricted-syntax
-      throw new HttpsError('invalid-argument', 'Invalid account ID');
-    }
-    const accountId = accountIdResult.value;
-
-    const {url} = request.data;
-
-    const logDetails = {url, accountId} as const;
-
-    logger.log(`[SUBSCRIBE] Subscribing account to feed source via URL...`, logDetails);
-
-    const subscribeToUrlResult = await rssFeedService.subscribeAccountToUrl({url, accountId});
-    if (!subscribeToUrlResult.success) {
-      logger.error(
-        prefixError(
-          subscribeToUrlResult.error,
-          '[SUBSCRIBE] Error subscribing account to feed source via URL'
-        ),
-        logDetails
-      );
-      // eslint-disable-next-line no-restricted-syntax
-      throw new HttpsError('internal', subscribeToUrlResult.error.message);
-    }
-
-    const userFeedSubscription = subscribeToUrlResult.value;
-
-    logger.log(`[SUBSCRIBE] Successfully subscribed account to feed source`, {
-      ...logDetails,
-      feedSourceId: userFeedSubscription.feedSourceId,
-      userFeedSubscriptionId: userFeedSubscription.userFeedSubscriptionId,
+    return subscribeAccountToFeedHelper({
+      auth: request.auth,
+      data: request.data,
+      rssFeedService,
     });
-
-    return {
-      userFeedSubscriptionId: userFeedSubscription.userFeedSubscriptionId,
-    };
   }
 );
 
@@ -310,46 +230,10 @@ export const subscribeAccountToFeedOnCall = onCall(
 export const createImportQueueItemOnFeedItemCreated = onDocumentCreated(
   `${FEED_ITEMS_DB_COLLECTION}/{feedItemId}`,
   async (event) => {
-    const feedItemId = parseFeedItemId(event.params.feedItemId);
-    if (!feedItemId.success) {
-      const betterError = prefixError(feedItemId.error, 'Failed to parse feed item ID');
-      logger.error(betterError, {feedItemId: event.params.feedItemId});
-      return;
-    }
-
-    const feedItemData = event.data?.data();
-    if (!feedItemData) {
-      logger.error(new Error('No feed item data found'), {feedItemId: feedItemId.value});
-      return;
-    }
-
-    const feedItemResult = parseFeedItem(feedItemData);
-    if (!feedItemResult.success) {
-      const betterError = prefixError(feedItemResult.error, 'Failed to parse feed item');
-      logger.error(betterError, {feedItemId: feedItemId.value, feedItemData});
-      return;
-    }
-
-    const feedItem = feedItemResult.value;
-
-    const createImportQueueItemResult = await importQueueService.create({
-      feedItemId: feedItem.feedItemId,
-      accountId: feedItem.accountId,
-      url: feedItem.url,
-    });
-
-    if (!createImportQueueItemResult.success) {
-      const betterError = prefixError(
-        createImportQueueItemResult.error,
-        'Failed to create import queue item'
-      );
-      logger.error(betterError, {feedItemId: feedItem.feedItemId});
-      return;
-    }
-
-    logger.warn('Successfully created import queue item', {
-      feedItemId: feedItem.feedItemId,
-      importQueueItemId: createImportQueueItemResult.value.importQueueItemId,
+    await createImportQueueItemHelper({
+      feedItemId: event.params.feedItemId,
+      data: event.data?.data(),
+      importQueueService,
     });
   }
 );
@@ -360,73 +244,11 @@ export const createImportQueueItemOnFeedItemCreated = onDocumentCreated(
 export const handleFeedUnsubscribeOnUpdate = onDocumentUpdated(
   `${USER_FEED_SUBSCRIPTIONS_DB_COLLECTION}/{userFeedSubscriptionId}`,
   async (event) => {
-    const {userFeedSubscriptionId: maybeUserFeedSubscriptionId} = event.params;
-
-    // Parse the user feed subscription ID.
-    const parseIdResult = parseUserFeedSubscriptionId(maybeUserFeedSubscriptionId);
-    if (!parseIdResult.success) {
-      logger.error(
-        prefixError(parseIdResult.error, '[UNSUBSCRIBE] Invalid user feed subscription ID'),
-        {maybeUserFeedSubscriptionId}
-      );
-      return;
-    }
-    const userFeedSubscriptionId = parseIdResult.value;
-
-    // Parse the before and after data.
-    const beforeData = event.data?.before.data();
-    const afterData = event.data?.after.data();
-
-    if (!beforeData || !afterData) {
-      logger.error(new Error('[UNSUBSCRIBE] Missing before or after data'), {
-        userFeedSubscriptionId,
-      });
-      return;
-    }
-
-    const beforeResult = parseUserFeedSubscription(beforeData);
-    const afterResult = parseUserFeedSubscription(afterData);
-
-    if (!beforeResult.success || !afterResult.success) {
-      logger.error(new Error('[UNSUBSCRIBE] Failed to parse user feed subscription data'), {
-        userFeedSubscriptionId,
-        beforeError: beforeResult.success ? undefined : beforeResult.error.message,
-        afterError: afterResult.success ? undefined : afterResult.error.message,
-      });
-      return;
-    }
-
-    const before = beforeResult.value;
-    const after = afterResult.value;
-
-    // Only proceed if the subscription was marked as inactive.
-    const becameInactive = before.isActive && !after.isActive;
-    if (!becameInactive) return;
-
-    const logDetails = {
-      userFeedSubscriptionId,
-      accountId: after.accountId,
-      feedSourceId: after.feedSourceId,
-      url: after.url,
-    } as const;
-
-    logger.log('[UNSUBSCRIBE] Processing unsubscribe for feed subscription', logDetails);
-
-    const unsubscribeResult = await rssFeedService.unsubscribeAccountFromUrl({
-      feedSourceId: after.feedSourceId,
-      url: after.url,
-      accountId: after.accountId,
+    await handleFeedUnsubscribeHelper({
+      userFeedSubscriptionId: event.params.userFeedSubscriptionId,
+      beforeData: event.data?.before.data(),
+      afterData: event.data?.after.data(),
+      rssFeedService,
     });
-
-    if (!unsubscribeResult.success) {
-      const betterError = prefixError(
-        unsubscribeResult.error,
-        '[UNSUBSCRIBE] Error unsubscribing account from Superfeedr feed'
-      );
-      logger.error(betterError, logDetails);
-      return;
-    }
-
-    logger.log('[UNSUBSCRIBE] Successfully unsubscribed account from Superfeedr feed', logDetails);
   }
 );
